+++ src/Hacks/ClassicParticles.cpp
@@ -9,7 +9,6 @@ class $modify (PlayerObject)
 	virtual void update(float p0)
 	{
 		PlayerObject::update(p0);
-
 		#ifdef GEODE_IS_ANDROID
 		auto particle = m_unk6e4;
 		#else